kernel: Use linux-yocto released version for qemu BSPs
Issue: LINCD-9205

We tried to use latest source to have patches in place to avoid fix them
locally. But it turns out to bring in more issues since the linux-yocto
maintainer sometimes updates the source code without enough build tests.

We might as well use released versions which is kind of behind but more stable.

Signed-off-by: He Zhe <zhe.he@windriver.com>