qemu: vcpu: avoid using ifdef CONFIG_KVM in common code
Issue: LINCD-9511

Add a new commit to deal with the fact that CONFIG_KVM is now not
allowed to be used in common code.  This leaves the original vcpu
commit largely intact as it was from the original author.

Three line stub addition lets us delete CONFIG_KVM from common code.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>