CGL: drop our in-house created userspace VRF package.
Issue: LINUXEXEC-23264

We have:

 DESCRIPTION="Based on Linux container and underlying namespace, \
 vrf package provides a framework to make easier to establish \
 virtual router/switch with WRLinux"

The vrf-1.1.tgz removed here doesn't let the reviewer peer into its
contents, but when one does, we see the AUTHORS file contains 1 line:

	yao.zhao@windriver.com

...indicating it was written in-house.  Further, an external search
shows only a reference to our own CGL product documentation (package
listing) - so it doesn't appear to have ever gathered any traction out
there in forming a user base, even though the file dates in the latest
tar are still from early 2015.  It is also a bit odd that even though
"we" are the authors/maintainers of this vrf package, we proceeded to
keep adding incremental patches to it (as seen in this commit diffstat).

If we look at the current official Linux VRF documentation:

https://docs.kernel.org/networking/vrf.html

we see that all the VRF configuration is done with the "ip" command,
which isn't surprising and probably explains a lot - if we go back to
the early days of CGL (2005-2008) we largely still had "ifconfig" users,
and use/knowledge of the "ip" command was still largely pending.  So
with that context one can somewhat understand why this was created.

Obviously we don't want to be maintaining old and/or alternate tooling
or wrappers for configuring things from userspace when the defacto
standard has been documented as such for years (see linux git history).

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>