wrlinux: move WRL_RECIPES updates out of templates
Issue: LINUXEXEC-23239

We put updates to WRL_RECIPES into templates as it made lots of
sense. Recipes would only be available when a template was in use and
the template ensured the environment the recipe would run in was
'complete'.

Another way to look at this, however, was the WRL_RECIPES was dynamic,
depending on how a build was configured. And this could be confusing
or inconvenient. Especially as some users use Wind River Linux like
the Yocto Project and thus avoid the use of templates.

Support of a recipe is binary in Wind River Linux, we either support
it or we don't. The use or non-use of template doesn't change this. By
moving updates to WRL_RECIPES out of templates the list remains
unchanged regardless of what, if any, templates are used.

Here we move all WRL_RECIPES updates from templates in the wrlinux
layer and all external layers, such as BSP layers into
conf/wrlinux-recipelists.conf. There are still some cases of
WRL_RECIPES being updated elsewhere, but not in templates, which is
still valid. For example a Wind River only layer can include
WRL_RECIPES in the layer.conf, making it a single repo operation to
add/remove recipes.

Since this change is extensive the opportunity was taken to do some
housekeeping as well, so some reordering is visible that doesn't
affect the outcome.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>