CGL: relocate template feature ipmi outside of wrlinux-cgl
Issue: LINUXEXEC-23264

We got an explicit request to keep IPMI during our feedback solicitation
and it really was the only feature I was explicitly considering at the
outset for a default move (vs. remove) anyway.

Note that the template.conf isn't really a "new file", but rather we
coalesced all the CGL configuration content together in an earlier
part of the wrlinux-cgl retirement.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>