recipelists: add ipmitool from what was wrlinux-cgl
Issue: LINCD-10656

As part of the demand driven restoration of packages that were previously a
part of wrlinux-cgl, this restores ipmitool to the supported recipe list.  Note
also that they are restored at a top level in keeping with the recipe changes
in LINUXEXEC-23239.

Note that currently ipmitool doesn't exist in any -dl layer, so NO_NETWORK
setting has to be commented out - however my understanding is that this was
a pre-existing condition the test team were aware of - as we did not
remove ipmitool package from wrlinux-cgl-dl layer.

In order to ensure we don't break world builds with the missing download, we
explicitly exclude it while discussions of adding it to a -dl take place.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>