templates: restore snmp from what was wrlinux-cgl
Issue: LINUXEXEC-23480

This restores the content of templates/feature/snmp as it existed
just prior to wrlinux-cgl removal.  With the difference that the
WRL_RECIPE line is removed - we already added it previously at the
top level as it was a dependency of openipmi.

We are restoring this as net-snmp is listed in a customer package
list table that can be found in case LINUXPG-1374

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>