recipelists: add frr, ifenslave and traceroute from what was wrlinux-cgl
Issue: LINUXEXEC-23480

As part of the demand driven restoration of packages that were
previously a part of wrlinux-cgl, this restores 3 packages to the
supported recipe list.  Note also that they are restored at a top level
in keeping with the recipe changes in LINUXEXEC-23239.

Currently ifenslave and frr don't exist in any -dl layer, so the
NO_NETWORK setting has to be commented out - however my understanding is
that this was a pre-existing condition the test team were aware of - as
we did not remove either package from the wrlinux-cgl-dl layer.

In order to ensure we don't break world builds with the missing
downloads, we explicitly exclude them while discussions of adding to
an existing -dl take place.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>