wrlinux: fixup a late addition to WRL_RECIPES
Issue: LINUXEXEC-23239

Wind River templates should not set WRL_RECIPES in most cases. The
container-registry template adds to WRL_RECIPES so we correct this
here. The 'skopeo' addition was a duplicate and can be dropped. The
'docker-distribution' addition is present in the
wrlinux-ovp-recipelists.conf file but we add it to
wrlinux-recipelists.conf to ensure this recipe can be used without
wrlinux-ovp.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>