Revert "boost: add boost 1.75.0"
Issue: LINUXEXEC-23512

This reverts commit f99b735cdb37b7d904ba99d8af4f62346bb72379.

wrlinux layer has higher priority than oe-core, adding a recipe
here overrides the one in oe-core regardless of the version.

We can see this low version boost recipe was added because vsomeip
needs it, and PREFERRED_VERSION is set in templates. But in fact,
the PREFERRED_VERSION setting has no effect.

Tracked-by: LINCD-10823

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>