feature/preserve-buildtime: add template to preserve buildtime
Issue: LINCD-11604

This template enables users to preserve buildtime. For now, it
contains two settings.

  KERNEL_DEBUG_TIMESTAMPS = "1"
  WRL_PRESERVE_BUILDTIME = "1"

KERNEL_DEBUG_TIMESTAMPS is used to preserve the kernel's build
date. The date can be examined via /proc/version or 'uname -a'
at runtime.

WRL_PRESERVE_BUILDTIME, on the other hand, causes SOURCE_DATE_EPOCH
to be set to the current time. In this way, the buildtime of a
package is preserved.

SOURCE_DATE_EPOCH, by default, is set to the date of the source
code git repo's latest commit of the current branch, or is set
to the date of some specific file such as Changelog. This is a
requirement for reproducible build. See the following links for
more info.

  https://reproducible-builds.org/docs/source-date-epoch/
  https://gcc.gnu.org/onlinedocs/cpp/Environment-Variables.html
  https://wiki.debian.org/ReproducibleBuilds/StandardEnvironmentVariables#Checklist
  https://wiki.yoctoproject.org/wiki/Reproducible_Builds

We use WRL_PRESERVE_BUILDTIME to change its default behavior.

Note that we use a template here to allow adding more settings
in the future if more use cases are identified.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>