wrlinux-ovp: delete main wrlinux-ovp configuration file
Issue: LINUXEXEC-23498

Aside from the DPDK version reference at the bottom, there are no useful
settings in this file we want to preserve.  And chances are the DPDK
values are stale and needing to be looked at anyways.

We may have had OVP users who didn't realize that the preempt-rt kernel
was auto-selected for them, so perhaps we want to mention that if/when
we make some kind of brief OVP migration guide.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>