wrlinux-ovp: delete all references to aufs.
Issue: LINUXEXEC-23498

This gets separate treatment for the following reasons:

The AUFS has always been problematic on -rt as it is an out-of-tree
filesystem, and the design doesn't "play" nice with in-tree integration,
nor has it been wholly inert when patched in but disabled.

I've had to fix many of these issues in the past, and if there ever was
a time to cut ties with AUFS, the removal of OVP is that time.

https://github.com/sfjro/aufs-standalone/issues/10

I'm hoping I can get Yocto to drop patching it into the kernel for 2023
but of course I can't guarantee success there.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>