wrlinux-ovp: import rt-tune and rt-latency
Issue: LINUXEXEC-23498

Copy across the configurations from wrlinux-ovp.  These are needed in
order for the test team to continue testing a use case we've decided we
need to continue to support beyond OVP.

Note that there was a naming inconsistency in the OVP files; in that
there was:

 guest-rt-tune
 host-rt-tune
 rt-latency-guest
 rt-latency-host

In this copy, we go with rt-<latency/tune>-<guest/host> for consistency.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>