wrlinux-ovp: delete glusterfs server filesystem stuff
Issue: LINUXEXEC-23498

I was originally going to give this a free pass and relocate the
packagegroup up one level based on input from others.

Move it now and then independently submit a removal later...

But when I did, I realized I'd have to create a whole new entry for
the layer meta-cloud-services in standard product.

So, rather than do and then undo all that, I've decided to go back to my
original position to just remove this and wait for demand to prove me
wrong in that decision.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>