wrlinux-ovp: delete already copied over rt-tune/rt-latency files
Issue: LINUXEXEC-23498

In commit 432e4d92d9ad ("wrlinux-ovp: import rt-tune and rt-latency")
we already copied these files across (instead of a "git mv") in order
to enable earlier transition by test team.

Now that we are following through with the rest of the removal, we can
remove the original copies.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>