wrlinux-ovp: relocate packagegroups that still may have value.
Issue: LINUXEXEC-23498

We may have use cases that drag these in.  We can revisit removal at a
later date.  There is no reason for them to hold up the bigger goal of
removing wrlinux-ovp as a whole.

When we have time, we can start by looking at criu as it is a server
oriented package and not directly in our target market at this time.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>