diod: remove from core-trace-tools
Issue: LINUXEXEC-23498

diod was used in OVP and somehow didn't get dropped from core-trace-tools.
There do not seem to be any package dependencies for it so it can be safely removed.
Drop it from the core-trace-tools until we can see if it's really needed and
if so, can be added back when things are stable.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>