preempt-rt: remove temp patch for i915 suspend/resume
Issue: LIN1023-2956

This change facilitated upstream discussion which eventually led to
a solution so it served its purpose.  But it will now conflict with
the upstream fix and result in one of those annoying "hot-patch"
failures.  Time to remove it.

The fix was for v6.4 but my backport to v6.1 has finally looped around.

https://lists.yoctoproject.org/g/linux-yocto/message/13392

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>